Abstracted task menu into SettingsMenu.vue
