+++ frontend/src/components/Task.vue
@@ -14,6 +14,7 @@
         </v-list-item-title>
         <v-text-field
           v-if="editing"
+          ref="taskNameInput"
           v-model="task.name"
           class="edit-task"
           filled
@@ -24,51 +25,12 @@
       </v-list-item-content>
 
       <v-list-item-action>
-        <v-menu offset-y>
-          <template v-slot:activator="{ on }">
-            <v-btn
-              icon
-              v-on="on"
-            >
-              <v-icon color="grey lighten-1">
-                mdi-information
-              </v-icon>
-            </v-btn>
-          </template>
-          <v-list>
-            <v-list-item two-line>
-              <v-list-item-content>
-                <v-list-item-title>{{ dateType }} on</v-list-item-title>
-                <v-list-item-subtitle>
-                  <span>{{ displayDate }} {{ displayTime }}</span>
-                </v-list-item-subtitle>
-              </v-list-item-content>
-            </v-list-item>
-          </v-list>
-          <v-divider />
-          <v-list>
-            <v-btn
-              color="warning"
-              fab
-              small
-              dark
-              class="menu-btn"
-              @click="editing = true"
-            >
-              <v-icon>mdi-pencil</v-icon>
-            </v-btn>
-            <v-btn
-              color="error"
-              fab
-              small
-              dark
-              class="menu-btn"
-              @click="deleteThisTask"
-            >
-              <v-icon>mdi-delete</v-icon>
-            </v-btn>
-          </v-list>
-        </v-menu>
+        <SettingsMenu
+          :date-type="dateType"
+          :date="date"
+          :edit-this="editThisTask"
+          :delete-this="deleteThisTask"
+        />
       </v-list-item-action>
     </v-list-item>
   </v-card>
@@ -76,10 +38,11 @@
 
 <script>
 import { mapMutations } from 'vuex'
-import moment from 'moment'
+import SettingsMenu from './SettingsMenu'
 
 export default {
   name: 'Task',
+  components: { SettingsMenu },
   props: {
     task: {
       type: Object,
@@ -103,12 +66,6 @@ export default {
     },
     date: function () {
       return this.task.completed ? this.task.completedDate : this.task.createdDate
-    },
-    displayDate: function () {
-      return moment(this.date).format('ddd MMM DD YYYY,')
-    },
-    displayTime: function () {
-      return moment(this.date).format('h:mm a')
     }
   },
   methods: {
@@ -116,6 +73,13 @@ export default {
       'completeTask',
       'deleteTask'
     ]),
+    editThisTask () {
+      this.editing = true
+      this.$nextTick()
+      this.$nextTick(function () {
+        this.$refs.taskNameInput.$el.querySelector('input').focus()
+      })
+    },
     completeThisTask () {
       this.completeTask({
         taskId: this.task.id,
@@ -133,7 +97,5 @@ export default {
 </script>
 
 <style scoped>
-  .menu-btn {
-    margin-left: 8px;
-  }
+
 </style>
